use u8 instead of usize

+++ analyze/src/lib.rs
@@ -20,7 +20,7 @@ pub enum KanjiNode {
         children: Vec<KanjiNode>,
     },
     Stroke {
-        index: usize,
+        index: u8,
         path: lyon_path::Path,
     },
 }
